perf(community): 게시글 조회 성능 최적화 (Redis 캐싱, Write-Back 및 DB 인덱싱 적용)
1. Redis 캐싱 전략 재설계 (Split & Merge)
   - 게시글 본문(변경 빈도 낮음)과 통계/댓글(변경 빈도 높음)을 분리하여 관리
   - 본문 정보는 post:info:{id} 키로 캐싱하여 DB 부하 감소
   - 댓글 목록은 캐싱하지 않고 DB에서 실시간 조회하여 데이터 정합성 보장

2. 조회수 Write-Back(쓰기 지연) 구현
   - 조회수 증가 요청 시 DB Update 없이 Redis에만 카운트 누적
   - PostBatchScheduler 추가: 10분 주기로 Redis의 누적 조회수를 DB에 일괄 반영
   - PostRepository: 대량 업데이트를 위한 incrementViewCount 쿼리 추가

3. DB 인덱스(Index) 최적화 적용
   - PostEntity: 카테고리별 조회순/최신순 정렬 및 유저별 조회 인덱스 추가 (idx_post_category_view, idx_post_category_date 등)
   - PostCommentEntity: 게시글별 댓글 목록 조회를 위한 복합 인덱스 추가 (idx_comment_post)
   - PostBookmarkEntity: 내 북마크 목록 최신순 조회를 위한 인덱스 추가 (idx_bookmark_user_date)

4. Service 및 Entity 리팩토링
   - PostEntity: 조회수, 댓글수, 북마크수 증감 편의 메서드 추가
   - PostService: 댓글 생성/삭제 시 게시글 본문 캐시를 불필요하게 삭제하지 않도록 개선

+++ community-service/src/main/java/com/example/communityservice/dto/response/PostDetailResponse.java
@@ -20,8 +20,8 @@ public class PostDetailResponse {
     private PostCategory category;
     private String title;
     private String content;
-    private int viewCount;
-    private int bookmarkCount;
+    private Long viewCount;
+    private Long bookmarkCount;
     private boolean isSolved;
     private boolean isBookmarked; // 로그인 유저의 북마크 여부
     private LocalDateTime createdAt;
